modify type annotation to return floats

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -6,7 +6,7 @@ import random
 from typing import AsyncIterator
 
 
-async def async_generator() -> AsyncIterator[int]:
+async def async_generator() -> AsyncIterator[float]:
     """
     Runs a loop 10 times, waiting 1 second each time, and yields a
     random number between 0 and 10
